Make the fully hosting configurable
