+++ src/Serenity/Hosting/ISerenityHosting.cs
@@ -1,0 +1,10 @@
+using FubuMVC.Core;
+
+namespace Serenity.Hosting
+{
+    public interface ISerenityHosting
+    {
+        IApplicationUnderTest Start(ApplicationSettings settings, FubuRuntime runtime, IBrowserLifecycle lifecycle);
+        void Shutdown();
+    }
+}
